% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{manhattan plot}
\usage{
manhattan(dataframe, title = NULL, max.y = "max", suggestiveline = 0,
  genomewideline = -log10(5e-08), size.x.labels = 9, size.y.labels = 10,
  annotate = F, SNPlist = NULL)
}
\arguments{
\item{dataframe, }{same as qqman package, 4 columns including SNP, CHR, BP and P}

\item{title}{title}

\item{max.y}{max}

\item{suggestiveline}{0}
}
\value{
genomewideline default -log10(5e-8)

size.x.labels default 9

size.y.labels default 10

annotate F

SNPlist NULL
}
\description{
manhattan plot
}
\details{
manhattan plot
}
\examples{
gwasResults <- data.frame(SNP=paste0("rs",1:40), CHR=1, BP = 1:40, P=runif(40))
manhattan(gwasResults)

}
\author{
Caleb
}
