% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayKS.R
\name{pathwayKS}
\alias{pathwayKS}
\title{pathway analysis}
\usage{
pathwayKS(wholePvalues, wholeName, fdr = 1.1, database = NULL,
  pathSizeMin = 15, pathSizeMax = 200)
}
\arguments{
\item{wholePvalues}{a vector of p values for all genes}

\item{wholeName}{a vector of all genes symbols}

\item{fdr}{desired FDR cutoff. Only pathways passed the FDR cutoff will survive}

\item{database}{a list of pathway database, each list element contains a vector of genes}
}
\value{
pathway enrichment result table
}
\description{
pathway enrichment analysis using KS test
}
\details{
pathway enrichment analysis using KS exaxt test
}
\examples{
wholePvalues <- rnorm(26)
wholeName <- letters[1:26]
database <- list(d1=letters[2:20],d2=letters[5:26])
pathwayKS(wholePvalues, wholeName, database=database, fdr=1)

}
\author{
Caleb
}
